# docs/swagger.yaml
definitions:
  oms.OrderForSwag:
    properties:
      account:
        type: string
      avg_px:
        type: string
      clord_id:
        type: string
      closed:
        type: string
      id:
        type: integer
      maturity_day:
        type: integer
      maturity_month_year:
        type: string
      open:
        type: string
      ord_type:
        type: string
      order_id:
        type: string
      price:
        type: string
      put_or_call:
        type: string
      quantity:
        type: string
      security_desc:
        type: string
      security_type:
        type: string
      session_id:
        type: string
      side:
        type: string
      stop_price:
        type: string
      strike_price:
        type: string
      symbol:
        type: string
      text:
        type: string
      time_in_force:
        type: string
    type: object
info:
  contact: {}
  termsOfService: http://somewhere.com/
  title: Traderui API
  version: "1.0"
paths:
  /orders:
    get:
      description: get all order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/oms.OrderForSwag'
            type: array
      summary: getOrders
    post:
      consumes:
      - application/json
      description: new 1 Order api
      parameters:
      - description: 'Order data for sending to executor '
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/oms.OrderForSwag'
      produces:
      - application/json
      responses:
        "202":
          description: OK
          schema:
            type: string
      summary: newOrder
  /orders/:id:
    delete:
      description: cancel 1 order
      parameters:
      - description: id of order
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/oms.OrderForSwag'
      summary: deleteOrder
    get:
      description: get 1 order
      parameters:
      - description: id of order
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oms.OrderForSwag'
      summary: getOrder
schemes:
- https
- http
swagger: "2.0"